GH130: Add user_id into link paths in profile view

+++ app/controllers/ProfileController.php
@@ -9,26 +9,25 @@ class ProfileController extends \Framework\AbstractController {
 		}
 	}
 
-	public function indexAction() {
+	public function indexAction($user_id = null) {
 		$session = $this->session->get('auth');
-		$user_id = $session->getUserCredentials()['id'];
-		$usersTable = new Users();
-		// Get user
-		$user = $usersTable->getUserByID($user_id);
-		$this->view->setVar('user', $user);
+		if ($user = $this->_getUserByID($user_id)) {
+			return $this->view->setVar('user', $user);
+		}
+		
 	}
 
-	public function usernameAction() {
+	public function usernameAction($user_id = null) {
+		$session = $this->session->get('auth');
+		// validate whether user_id is md5 value and get user
+		if (!$user = $this->_getUserByID($user_id)) return false;
 		// In this action implemented change username function
-		if (!$this->view->form) $this->view->setVar('form', new \ProfileForm\UsernameForm());
+		if (!$this->view->form) $this->view->setVar('form', new \ProfileForm\UsernameForm($user));
 		if ($this->getDI()->getRequest()->isPost()) {
 			$login = (Object) Array();
 			if ($this->view->form->isValid($this->getDI()->getRequest()->getPost(), $login)) {
-				$session = $this->session->get('auth');
-				$usersTable = new Users();
-				$user_id = $session->getUserCredentials()['id'];
-				$usersTable->getUserByID($user_id)->setNewUsername($login);
-				return $this->response->redirect('profile/index');
+				$user->setNewUsername($login);
+				return $this->dispatcher->forward(array('controller' => 'profile','action' => 'index' ));
 			}
 			else {
 				$this->view->messages->addError('Please, fix errors and try again!');
@@ -36,15 +35,15 @@ class ProfileController extends \Framework\AbstractController {
 		}
 	}
 
-	public function passwordAction() {
-		if (!$this->view->form) $this->view->setVar('form', new \ProfileForm\PasswordForm());
+	public function passwordAction($user_id = null) {
+		$session = $this->session->get('auth');
+		// validate whether user_id is md5 value and get user
+		if (!$user = $this->_getUserByID($user_id)) return false;
+		if (!$this->view->form) $this->view->setVar('form', new \ProfileForm\PasswordForm($user));
 		if ($this->getDI()->getRequest()->isPost()) {
 			$password = (Object) Array();
 			if ($this->view->form->isValid($this->getDI()->getRequest()->getPost(), $password)) {
-				$session = $this->session->get('auth');
-				$user_id = $session->getUserCredentials()['id'];
-				$usersTable = new Users();
-				$passwordChanged = $usersTable->getUserByID($user_id)->changeUserPassword($password);
+				$passwordChanged = $user->changeUserPassword($password);
 				// redirect to sing up confirmation page
 				if ($passwordChanged) return $this->view->pick('profile/password_confirmation');
 			}
@@ -57,20 +56,24 @@ class ProfileController extends \Framework\AbstractController {
 		}
 	}
 
-	public function emailAction() {
-		if (!$this->view->form) $this->view->setVar('form', new \ProfileForm\EmailForm());
+	public function emailAction($user_id = null) {
+		$session = $this->session->get('auth');
+		// validate whether user_id is md5 value and get user
+		if (!$user = $this->_getUserByID($user_id)) return false;
+		if (!$this->view->form) $this->view->setVar('form', new \ProfileForm\EmailForm($user));
 		if ($this->getDI()->getRequest()->isPost()) {
-			$email = (Object) Array('user_id' =>'','is_primary'=>''); // pre-populate required data
+			// pre-populate required data
+			$email = (Object) Array(
+				'user_id' => $user_id, 'is_primary' => ''
+			);
 			if ($this->view->form->isValid($this->getDI()->getRequest()->getPost(), $email)) {
-				$session = $this->session->get('auth');
-				$email->user_id = $session->getUserCredentials()['id'];
 				$email->is_primary = 0; // not primary
 				$usersEmails = new UsersEmails();
 				$config = $this->getDI()->get('config');
 				$usersEmails->createEmail($email);
 				$usersEmails->sendVerifyEmail($config);
 				// redirect to profile/index
-				if ($usersEmails) return $this->response->redirect('profile/index');
+				if ($usersEmails) return $this->dispatcher->forward(array('controller' => 'profile','action' => 'index' ));
 			} else {
 				// output error message
 				$this->view->messages->addError('Please, fix errors and try again!');
@@ -78,12 +81,14 @@ class ProfileController extends \Framework\AbstractController {
 		}
 	}
 
-	public function deleteemailAction() {
+	public function deleteemailAction($user_id = null) {
+		$session = $this->session->get('auth');
+		// validate whether user_id is md5 value and get user
+		if (!$user = $this->_getUserByID($user_id)) return false;
 		if (!$this->view->form) $this->view->setVar('form', new \ProfileForm\DeleteEmailForm());
-		$deleteEmail = (Object) Array();
+		// pre-populate required data
+		$deleteEmail = (Object) Array('user_id' => $user_id);
 		if ($emailDeleted = $this->view->form->isValid($this->getDI()->getRequest()->getQuery(), $deleteEmail)) {
-			$session = $this->session->get('auth');
-			$deleteEmail->user_id = $session->getUserCredentials()['id'];  // pre-populate required data
 			$usersEmails = new UsersEmails();
 			$userEmail = $usersEmails->getEmailByIDandUserID($deleteEmail);
 			// Email should exist and it should not be primary. Without primary email user could not login to the system!
@@ -98,21 +103,35 @@ class ProfileController extends \Framework\AbstractController {
 				"action" => "index" ));
 	}
 
-	public function setprimaryemailAction() {
+	public function setprimaryemailAction($user_id = null) {
+		$session = $this->session->get('auth');
+		// validate whether user_id is md5 value and get user
+		if (!$user = $this->_getUserByID($user_id)) return false;
 		if (!$this->view->form) $this->view->setVar('form', new \ProfileForm\SetEmailAsPrimaryForm());
-		$primaryEmail = (Object) Array();
+		// pre-populate required data
+		$primaryEmail = (Object) Array('user_id' => $user_id);
 		if ($this->view->form->isValid($this->getDI()->getRequest()->getQuery(), $primaryEmail)) {
-			$session = $this->session->get('auth');
-			$primaryEmail->user_id = $session->getUserCredentials()['id']; // pre-populate required data
 			$usersEmails = new UsersEmails();
 			$usersEmails->setNewPrimaryEmail($primaryEmail);
 			// update session object
-			$usersTable = new Users();
-			$user = $usersTable->getUserByID($primaryEmail->user_id);
-			$this->session->set('auth', new \Auth($user));
+			if ($primaryEmail->user_id == $session->getUserId()) {
+				$this->session->set('auth', new \Auth($user));
+			}
 			// redirect to profile/index
-			if ($usersEmails) return $this->response->redirect('profile/index');
+			if ($usersEmails) return $this->dispatcher->forward(array('controller' => 'profile','action' => 'index' ));
 		}
+	}
+
+	protected function _getUserByID($user_id) {
+		$usersTable = new \Users();
+		if (\Framework\Validation\Validator\Md5::isValid($user_id) 
+			&& $user = $usersTable->getUserByID($user_id)) {
+			return $user;
+		}
+		$this->dispatcher->forward(
+			array('controller' => 'index','action' => 'index' )
+		);
+		return null;
 	}
 
 }
